Shader reflection is done in the new shader class.
* Shader reflection which was previously done per shader source, is now done on the Shader class.
* It uses the SPIR-V source to perform reflection. 
* In the future when we've fully transitioned to HLSL 2021, we can remove the SPIR-V Cross and SPIR-V Tools libraries.
