+++ Engine/XenonBackend/CMakeLists.txt
@@ -43,6 +43,7 @@ set(
 	"BottomLevelAccelerationStructure.hpp"
 	"TopLevelAccelerationStructure.hpp"
 	"RayTracingPipeline.hpp"
+	"Shader.cpp"
 	"Shader.hpp"
 )
