+++ Engine/XenonDX12Backend/DX12ComputePipeline.cpp
@@ -82,7 +82,7 @@ namespace /* anonymous */
 		XENON_DX12_ASSERT(pReflector->GetDesc(&shaderDesc), "Failed to get the reflection description!");
 
 		// Setup resources.
-		for (const auto& resource : shader.getSPIRV().getResources())
+		for (const auto& resource : shader.getResources())
 		{
 			auto& binding = bindingInfos.emplace_back();
 			binding.m_Type = resource.m_Type;
